pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["Benjamin Cretois <benjamin.cretois@nina.no>"]
description = ""
name = "schoolmap"
package-mode = false
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
PyYAML = "6.0.1"
albumentations = "^1.4.18"
backoff = "^2.2.1"
# Prebulit wheel https://github.com/facebookresearch/detectron2/discussions/5200
detectron2 = {version = "0.6+2a420edpt2.3.1cu121", source = "torch_packages_builder"}
einops = "0.8.0"
geopandas = "1.0.1"
gradio = "^5.0.0"
lightning = "^2.4.0"
marimo = "^0.9.10"
matplotlib = "^3.9.0"
numpy = "^2.0.0"
opencv-python = "^4.10.0.84"
optuna = "^4.0.0"
python = "~=3.10.0"
rasterio = "1.4.1"
scikit-learn = "^1.5.2"
shapely = "^2.0.6"
tensorboard = "2.17.0"
tensorboard-data-server = "0.7.2"
torch = "2.3.1"
torchmetrics = "1.4.0.post0"
torchvision = "0.18.1"
tqdm = "4.66.4"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[[tool.poetry.source]]
name = "torch_packages_builder"
priority = "supplemental"
url = "https://miropsota.github.io/torch_packages_builder"
